requirements.txt
bert-tensorflow==1.0.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
tensorboard==1.13.1